% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bocaccio_recruitment_mcmc}
\alias{bocaccio_recruitment_mcmc}
\title{Bocaccio recruitment MCMC estimates (millions of age-1 fish) for
coastwide stock from the 2024 stock assessment}
\format{
A tibble, also of class \code{recruitea_recruitment} for plotting, with each row representing an MCMC sample, and columns:
\describe{
\item{\code{1935}:\code{2023}:}{year for each recruitment}
}
}
\source{
Generated from running \code{data-raw/bocaccio.R}
}
\usage{
bocaccio_recruitment_mcmc
}
\description{
As reported in the Science Response from the 2024 meeting; in press. Values
obtained from Rowan Haigh, 8th April 2024. TODO - get unfished equilibrium value.
}
\examples{
\dontrun{
bocaccio_recruitment_mcmc
plot(bocaccio_recruitment_mcmc)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
