% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitea-recruitment.R
\name{plot.recruitea_recruitment}
\alias{plot.recruitea_recruitment}
\title{Plot a recruitea recruitment time series object (currently assumes annual values)}
\usage{
\method{plot}{recruitea_recruitment}(
  obj,
  xlab = "Year",
  ylab = attr(obj, "axis_name"),
  subsample = 10,
  unfished_years_before = 3,
  y_tick_by = 1,
  x_tick_extra_years = 20,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  y_max = NULL,
  ...
)
}
\arguments{
\item{obj}{a \code{recruitea_recruitment} object, which is a tibble with rows
representing MCMC samples, the \code{unfished} column being the unfished
equilibrium, and remaining columns whose names are the years.}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis, defaults to a \code{axis-name} attribute of \code{obj}}

\item{subsample}{number of MCMC samples to plot}

\item{unfished_years_before}{years before the start to plot the unfished
equilibrium recruitment}

\item{y_tick_by}{increments for y tick marks}

\item{start_decade_ticks}{decade to start decadal tickmarks}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying \code{ylim})}

\item{...}{further options passed onto \code{plot.default()}}
}
\value{
plot of the time series with \code{subsample} number of MCMC samples to
the current device; returns nothing.
}
\description{
Temporal plot for a \code{recruitea} recruitment time series (of class
\code{recruitea_recruitment}) object. Default is to show multiple MCMC samples.
TODO Add an option to create pacea type recruitment plot.
}
\examples{
\dontrun{
plot(hake_recruitment_mcmc)
plot(hake_recruitment_mcmc,
     xlim = c(lubridate::dmy(01011950),
              lubridate::dmy(01012040))) # to expand x-axis
}
}
\author{
Andrew Edwards
}
