% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_recruitment_mcmc}
\alias{hake_recruitment_mcmc}
\title{Hake age-0 recruitment MCMC estimates from the 2024 assessment as another example}
\format{
A tibble with each row representing an MCMC sample, and columns:
\describe{
\item{unfished:}{presumably gives \code{R0}}
\item{\code{1966}:\code{2027}:}{year for each recruitment}
}
}
\source{
Generated from running \code{data-raw/hake.R}
}
\usage{
hake_recruitment_mcmc
}
\description{
Hake age-0 recruitment MCMC estimates from the 2024 assessment as another example
}
\examples{
\dontrun{
hake_recruitment_mcmc
plot(hake_recruitment_mcmc)  # when plot function written
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
